SearchResult

+++ AltexTravel.API/AltexTravel.API/AltexTravel.API.Amadeus/Models/SearchResult/PricePerAdult.cs
@@ -5,9 +5,9 @@ namespace AltexTravel.API.Amadeus.Models.SearchResult
     public class PricePerAdult
     {
         [JsonProperty("total")]
-        public double Total { get; set; }
+        public decimal Total { get; set; }
 
         [JsonProperty("totalTaxes")]
-        public double TotalTaxes { get; set; }
+        public decimal TotalTaxes { get; set; }
     }
 }
